Initial commit with PageRank Nibble implemented
We implemented the push, approximate PageRank, and PageRank Nibble algorithm of Andersen, Chung, and Lang [1] as well as methods to compute the network community profile (NCP) plot [2] as well as basic graph drawing using matplotlib and graphviz.

We need to clean up the code to allow command line inputs, and then look at other spectral methods for graph partitioning.

Reference:
[1] R. Andersen, F. Chung, and K. Lang, Using PageRank to Locally Partition a Graph
[2] Stanford Network Analysis Platform (SNAP) http://snap.stanford.edu/snap/index.html